refactor: change modalWindow location, make close element on card
