+++ react-components/src/components/ModalWindow/index.scss
@@ -16,12 +16,13 @@
   }
 
   &__close {
-    background-image: url(../../../assets/x.svg);
+    background-image: url(../../assets/x.svg);
     background-repeat: no-repeat;
     width: 40px;
     height: 40px;
     position: absolute;
-    top: -50px;
+    top: 5px;
+    right: 10px;
     opacity: 0.7;
     cursor: pointer;
   }
