+++ react-components/src/pages/Main/index.tsx
@@ -4,7 +4,7 @@ import Cards from '../../components/Cards';
 import Pagination from '../../components/Pagination';
 import { Card as ICard } from '../../types/card';
 import { useSearchParams } from 'react-router-dom';
-import ModalWindow from '../../components/Cards/ModalWindow';
+import ModalWindow from '../../components/ModalWindow';
 import Loader from '../../components/Loader';
 
 const Main = () => {
